+++ CHANGELOG.md
@@ -4,9 +4,10 @@ seleniumtid changelog
 v0.7
 ----
 
-*Release date: 2015-01-??*
+*Release date: 2014-12-23*
 
   * Allow to autocomplete self.driver and self.utils in IDEs
+  * Remove non-mandatory requirements
 
 v0.6
 ----